Fixed GPU memory leak & fixed crashes - Fixed GPU memory leak during training   - This was due to torchmetrics stateful class BinaryAveragePrecision not releasing memory   - Changing the code to use the functional equivalent fixed the issue   - Added manual del of heavy variables to ensure memory remains constant across training steps   - Added torch.no_grad to model testing to reduce GPU memory usage significantly - Fixed data loading not working with GPU   - Had to set num_workers=0 - Fixed collate_fn crashing due to pin_memory not being supported for tensor class - Fixed model loading during model testing - Changed target type from long to int8 - Fixed training not working when no logger is pecified - Fixed cometml logger not being created - Added parameters logging during training

+++ src/datasets/wildfire_data.py
@@ -16,7 +16,7 @@ class WildfireData:
         if len(dataset) == 0:
             return cls(
                 images=torch.empty((0, 0, 0), dtype=torch.float32),
-                masks=torch.empty((0, 0, 0), dtype=torch.long),
+                masks=torch.empty((0, 0, 0), dtype=torch.int8),
                 geotransforms=torch.empty((0, 0), dtype=torch.float32),
                 batch_size=[0],
             )
@@ -30,7 +30,7 @@ class WildfireData:
                 dtype=torch.float32,
             ),
             masks=MemoryMappedTensor.empty(
-                (len(dataset), *dataset[0][1].shape), dtype=torch.long
+                (len(dataset), *dataset[0][1].shape), dtype=torch.int8
             ),
             geotransforms=MemoryMappedTensor.empty(
                 (len(dataset), *dataset[0][2].shape), dtype=torch.float32
